Move usage segment to own component
