+++ src/app/modules/it-systems/it-system-usages/it-system-usage-details/it-system-usage-details-archiving/it-system-usage-details-archiving-usage/it-system-usage-details-archiving-usage.component.scss
@@ -1,8 +1,9 @@
-@import "theme";
-@import "mixins";
-
 :host {
   display: flex;
   flex-direction: column;
   width: 100%;
+}
+
+.notes-column-span {
+  grid-column: span 3;
 }
