+++ src/app/modules/it-systems/it-systems.module.ts
@@ -56,6 +56,7 @@ import { ItSystemInterfacesTableComponent } from './shared/it-system-interfaces-
 import { ItSystemKleOverviewComponent } from './shared/it-system-kle-overview/it-system-kle-overview.component';
 import { KleTableComponent } from './shared/kle-table/kle-table.component';
 import { ItSystemUsageDetailsArchivingCatalogComponent } from './it-system-usages/it-system-usage-details/it-system-usage-details-archiving/it-system-usage-details-archiving-catalog/it-system-usage-details-archiving-catalog.component';
+import { ItSystemUsageDetailsArchivingUsageComponent } from './it-system-usages/it-system-usage-details/it-system-usage-details-archiving/it-system-usage-details-archiving-usage/it-system-usage-details-archiving-usage.component';
 
 @NgModule({
   declarations: [
@@ -112,6 +113,7 @@ import { ItSystemUsageDetailsArchivingCatalogComponent } from './it-system-usage
     GdprOverviewComponent,
     GdprOverviewDialogComponent,
     ItSystemUsageDetailsArchivingCatalogComponent,
+    ItSystemUsageDetailsArchivingUsageComponent,
   ],
   imports: [CommonModule, SharedModule, ITSystemsRouterModule, MatInputModule, MatProgressSpinnerModule],
 })
